requirements.txt

# django
django==2.2.8

# django-bootstrap3
django-bootstrap3 ==12.0.3

# 图片验证码
django-simple-captcha==0.5.12

# SQL同步
PyMySQL==0.9.3
paramiko==2.4.2
cryptography==2.4.2

# websocket实时同步
dwebsocket==0.5.12